requirements.txt
python-gitlab
GitPython
